requirements.txt
Pillow==6.1.0
pytesseract==0.2.7
